8 money salary management automatization (#18)
* salary monitoring and wages issue

* com, dont know what

* hotfix

* calcultion of DailyWages added and much more

* DailyWage fixed and polished

+++ Core/Services/MasterServicesService.cs
@@ -86,7 +86,7 @@ public class MasterServicesService : IMasterServicesService
         _masterServicesRepository.Save();
     }
 
-    public void AddService(string name, decimal price)
+    public void AddService(string name, float price)
     {
         var service = new ProvidedService
         {
