+++ Infrastructure/Entities/ProvidedService.cs
@@ -7,6 +7,5 @@ public class ProvidedService : IEntity
     [Column("IdProvidedService")]
     public int Id { get; set; }
     public required string ServiceName { get; set; }
-    public decimal Price { get; set; }
-
+    public float Price { get; set; }
 }
